Accessing contents of the multiverse object

A multiverse object contains several Object variables. These can be accessed using convenient functions. Variables from the analysis that is being performed within the multiverse can be accessed using the $. Object variables such as the code, the expanded parameter options table, the parameters and the conditions can be accessed using respective functions

Usage

# S3 method for class 'multiverse'
multiverse$name

# S3 method for class 'multiverse'
multiverse$name <- value

expand(multiverse)

# Default S3 method
expand(multiverse)

# S3 method for class 'multiverse'
expand(multiverse)

size(multiverse)

# Default S3 method
size(multiverse)

# S3 method for class 'multiverse'
size(multiverse)

code(multiverse, .pretty = TRUE)

# Default S3 method
code(multiverse, .pretty = TRUE)

# S3 method for class 'multiverse'
code(multiverse, .pretty = TRUE)

parameters(multiverse)

# Default S3 method
parameters(multiverse)

# S3 method for class 'multiverse'
parameters(multiverse)

conditions(multiverse)

# Default S3 method
conditions(multiverse)

# S3 method for class 'multiverse'
conditions(multiverse)

extract_variable_from_universe(multiverse, idx, name)

Arguments

multiverse

Object of class multiverse

name

a variable name

value

a new value to be assigned

.pretty

A binary argument whether `code()` should prettify the output using the tidyverse style guide. defaults to TRUE.

idx

index of the universe in the multiverse (corresponds to the row in the table)