Survey to study the effect of fertility on religiosity and political attitudes

A dataset for the study conducted by Jung et al. (2014) in their study, doi: 10.1073/pnas.1402786111 Female hurricanes are deadlier than male hurricanes.

Format

A data frame with 502 rows and 26 variables:

Year

Year in which the hurricane occurred

Name

Name given to the hurricane

MasFem

A score of how Masculine or Feminine a hurricane is, on a scale of 1 - 11 where 1 is the most masculine and 11 is the most feminine. Each hurricane was rated by 9 independent coders

MinPressure_before

The minimum pressure of hurricanes at the time of landfall in the United States obtained from NOAA (www.aoml.noaa.gov/hrd/hurdat/All_U.S._Hurricanes.html)

Minpressure_Updated_2014

The minimum pressure of hurricanes at the time of landfall in the United States obtained from NOAA (www.aoml.noaa.gov/hrd/hurdat/All_U.S._Hurricanes.html)

Gender_MF

"Marriage is between a man and a woman."

Category

Category labels on a scale of 1 - 5, with 5 being the most severe or extreme

alldeaths

Total number of deaths caused by the hurricane. Information on death tolls of hurricanes were obtained primarily from monthly weather reports in the digital archive of the National Oceanic and Atmospheric Administration (www.aoml.noaa.gov/hrd/hurdat/mwr_pdf/)

HighestWindSpeed

The maximum wind speed of hurricanes at the time of landfall in the United States obtained from NOAA (www.aoml.noaa.gov/hrd/hurdat/All_U.S._Hurricanes.html). This data is only available for storms after 1979

NDAM

normalized damage (in million $)

Source

Source from where the data was gathered

Elapsed.Yrs

Time since hurricane

Details

This dataset was collated by Jung et. al in their study Female hurricanes are deadlier than male hurricanes They hypothesised that hurricanes with more feminine names might be perceived as less dangerous and hence lead to people taking fewer precautionary measures, resulting in more death and damages.

References

Kiju Jung and Sharon Shavitta and Madhu Viswanathana and Joseph M. Hilbed. (2014). "Female hurricanes are deadlier than male hurricanes." *Proceedings of the National Academy of Sciences*, 111(24), 8782-8787.

Yang Liu and Alex Kale and Tim Althoff and Jeff Heer. (2020). Boba: Authoring and Visualizing Multiverse Analyzes. *arXiv preprint* arXiv: 2007.05551 .