Survey to study the effect of fertility on religiosity and political attitudes

A dataset containing the responses to the survey coduncted by Durante, Rae, and Griskevicius (2013) in their study, The fluctuating female vote: Politics, religion, and the ovulatory cycle. Durante et al. study effect of fertility on religiosity and political attitudes. Steegen, Tuerlinckx, Gelman, and Vanpaemel (2016) used this dataset in their analysis to illustrate how a multiverse analysis can highlight the robustness of the conclusions reached by the original author.

durante

Format

A data frame with 502 rows and 26 variables:

Abortion

"Abortion is a women’s [sic] right."

DateTesting

Date of participant filling in the questionnaire.

Donate

"For the next part of the study we will donate $1 to the presidential campaign of your preferred candidate. Please indicate which candidate’s campaign you would like us to donate $1 to.” Mitt Romney — Barack Obama"

FreeMarket

"In nearly every instance, the free market allocates resources most efficiently."

Marijuana

"Marijuana should be legal."

Marriage

"Marriage is between a man and a woman."

PrivSocialSec

"Privatize Social Security." 1 – 7

Profit

"Business corporations make too much profit."

Rel1

"How much do you believe in God?"

Rel2

"I see myself as a religiously oriented person."

Rel3

"I believe that God or a Higher Power is responsible for my existence"

Relationship

What is your current romantic relationship status?” (1) not dating/romantically involved with anyone, (2) dating or involved with only one partner, (3) engaged or living with my partner, (4) married, or (5) other. If participants picked response (5), they were prompted to provide a description of their relationship, which was subsequently coded into one of the four options by the original authors. The data here has already been coded into another response option.

ReportedCycleLength

How many days long are your menstrual cycles? (for most women, the range is between 25-35 days) Keep in mind this is the number of days from the start of one menstrual period to the start of the next menstrual period and NOT the length of your menstrual bleeding.

RestrictAbortion

"Laws should restrict abortion in all or most cases."

RichTax

"The rich should pay a higher tax rate than the middle class."

StLiving

"Government should ensure that all citizens meet a certain minimum standard of living"

StartDateNext

Indicates the expected start date of their next menstrual period (the research material does not contain a question about the variable. However, the data file for Study 2 contained this variable.)

StartDateofLastPeriod

Please give your best estimate of the date on which you started your last period (please be as precise as possible). This date was probably within the last few weeks. Sometimes thinking of where you were when you started your last period helps. For instance, was it on a weekend?, were you at work, was it during a football game?, etc. Please write the date in mm/dd/yyyy format (e.g., 8/18/2012).

StartDateofPeriodBeforeLast

Please give your best estimate of the date on which you started the period before your last period (please be as precise as possible). Please write the date in mm/dd/yyyy format (e.g., 7/18/2012).”

StemCell

"Stem cell research is moral and can be useful for science."

Sure1

"How sure are you about that date (StartDateofLastPeriod)?"

Sure2

How sure are you about that date (StartDateofPeriodBeforeLast)?

Vote

"Imagine walking into the voting booth today. Who would you vote for in the presidential election?” Mitt Romney (republican) – Barack Obama (democrat)"

WorkerID

ID of participant

Details

All questions were preceded by the prompt --- "Please indicate how much you agree with the following statements"

The following items were responses to religiosity items (on a scale of 1 - 9): Rel1, Rel2, Rel3

The following items were responses to fiscal political attitudes items (on a scale of 1 - 7): RichTax, TooMuchProfit, StandardLiving, FreeMarket, PrivSocialSec

The following items were responses to social political attitudes items (on a scale of 1 - 7): Abortion, Marriage, StemCell, Marijuana, RestrictAbortion