Execute parts of, or the entire multiverse

These are functions which allow the user to execute parts or whole of the multiverse. The user can choose to either execute the default analysis using the execute_universe, or a part or whole of the multiverse using the execute_multiverse.

execute_multiverse(multiverse, cores = getOption("mc.cores", 1L))

execute_universe(multiverse, .universe = 1)

Arguments

multiverse
The multiverse object

cores
Indicates the number of cores to use. This will execute the entire multiverse in parallel. Defaults to NULL (running in a single core)

.universe
Indicate which universe to execute, if the user wants to execute a specific combination of the parameters using `execute_universe`. Defaults to NULL, which will execute the first (default) analysis.

Details

Each single analysis within the multiverse lives in a separate environment. We provide convenient functions to access the results for the default analysis, as well as parts or whole of the multiverse. Each analysis can also be accessed from the multiverse table, under the results column.

Examples

if (FALSE) {
#' M <- multiverse()
inside(M, {
  data <- rnorm(100, 50, 20)

  x.mean <- mean(data, trim = branch(
    trim_values,
    "trim_none" ~ 0,
    "trim_1pc" ~ 0.05,
    "trim_5pc" ~ 0.025,
    "trim_10pc" ~ 0.05
  ))
})

# Computes the analysis for all
# universes in the multiverse`
M %>%
  execute_multiverse()
}