Exporting results from a multiverse analysis to JSON

Exports the results of the multiverse analysis to JSON in a format which is compatible with the multiverse visualisation tool

Usage

export_2_json(x, term, mean, sd, dist, filename)

export_dist_2_json(x, term, dist, filename)

export_code_json(x, filename)

Arguments

x

a tidy tibble or data frame which contains summary statistics or distributional information of each regression coefficient parameter

term

column in the data frame, x, which contains the name of the coefficients

mean

column in the data frame, x, which contains the mean estimate for each coefficient

sd

column in the data frame, x, which contains the standard error estimate for each coefficient

dist

column in the data frame, x, which contains vectorised distributions---an object of class `distribution` for each coefficient

filename

filename on disk (as a character string)

Value

a data frame or a JSON file

Examples

# \donttest{
library(dplyr)
library(tidyr)
#> 
#> Attaching package: ‘tidyr’
#> The following object is masked from ‘package:multiverse’:
#> 
#>     expand

M = multiverse()

inside(M, {
  df = tibble(
    x = rnorm(100),
    y = x * 0.5 + rnorm(100, 0, 2)
 )
 
 # alternatives to remove outlier
 df.filtered = df %>%
   filter(
     branch(outlier_exclusion,
        "2SD" ~ abs(y - mean(y)) > 2*sd(y),
        "3SD" ~ abs(y - mean(y)) > 3*sd(y)
     )
 )
  
  fit = lm(y ~ x, data = df)
  res = broom::tidy(fit)
})

execute_multiverse(M)

multiverse::expand(M) %>%
  extract_variables(res) %>%
  unnest(res) %>%
  export_2_json(term, estimate, std.error) 
#> # A tibble: 2 × 3
#>   .universe outlier_exclusion results           
#>       <int> <chr>             <list>            
#> 1         1 2SD               <gropd_df [2 × 7]>
#> 2         2 3SD               <gropd_df [2 × 7]>
# }