Code corresponding to a single analysis

Given a particular set of options for each parameter, extracts the code for performing a single analysis from the code used to declare the multiverse. This function is called automatically and not exported.

get_code(.code, .assgn = NULL)

Arguments

.code

Code that is passed to the multiverse. This is not stripped of calls such as branch_assert().

.assgn

A list containing the assignments for each defined parameter in the multiverse

Details

For a particular parameter assignment (i.e. one set of options that each defined parameter in the multiverse takes), this function rewrites the code passed into the multiverse to output the corresponding code for that set of parameter values --- the analysis for a single universe.

This is primarily going to be called by other functions, and perhaps not going to be as useful to the user for anything other than inspecting the rewritten code.