Pass code into the multiverse

Add code to the multiverse using the function using a function call, or an assignment operator, which is a wrapper around the function

Usage

inside(multiverse, .expr, .label = NULL, .execute_default = TRUE)

Arguments

multiverse

A multiverse object. A multiverse object is an S3 object which can be defined using multiverse()

.expr

R syntax. All the operations that the user wants to perform within the multiverse can be passed. Since it accepts a single argument, chunks of code can be passed using \{\}. See example for details.

.label

It is extracted automatically from the code block of type multiverse when run in an RMarkdown document. This should be used only within an RMarkdown document. Defaults to NULL.

.execute_default

Should the default multiverse be executed as part of this call?

Value

a multiverse object

Details

The inside function can only access variables which can be accessed at the same environment where the multiverse object was declared in.

To perform a multiverse analysis, we will need to write code to be executed within the multiverse. The inside() functions allows us to do this. Use inside() to pass any code to the specified multiverse, which is captured as an expression. To define multiple analysis options in the code passed to the multiverse, use the branch() function.

See also

Examples

# \donttest{
M.1 <- multiverse()

# using `inside` to declare multiverse code
inside(M.1, {
  data <- rnorm(100, 50, 20)

  x.mean <- mean(data, trim = branch(
    trim_values,
    "trim_none" ~ 0,
    "trim_1pc" ~ 0.05,
    "trim_5pc" ~ 0.025,
    "trim_10pc" ~ 0.05
  ))
})

# declaring multiple options for a data processing step (calculating a new variable)
data(durante)
df <- durante

inside(M.1, {
  df <- df  %>%
    mutate( ComputedCycleLength = StartDateofLastPeriod - StartDateofPeriodBeforeLast ) %>%
    mutate( NextMenstrualOnset = branch(menstrual_calculation,
                                   "mc_option1" ~ StartDateofLastPeriod + ComputedCycleLength,
                                   "mc_option2" ~ StartDateofLastPeriod + ReportedCycleLength,
                                   "mc_option3" ~ StartDateNext
  ))
})

# }