'Explorable Multiverse' Data Analysis and Reports in R

Implement 'multiverse' style analyses (Steegen S., Tuerlinckx F, Gelman A., Vanpaemal, W., 2016) doi:10.1177/1745691616658637 to show the robustness of statistical inference. 'Multiverse analysis' is a philosophy of statistical reporting where paper authors report the outcomes of many different statistical analyses in order to show how fragile or robust their findings are. The 'multiverse' package (Sarma A., Kale A., Moon M., Taback N., Chevalier F., Hullman J., Kay M., 2021) doi:10.31219/osf.io/yfbwm allows users to concisely and flexibly implement 'multiverse-style' analysis, which involve declaring alternate ways of performing an analysis step, in R and R Notebooks. Users can also create 'Explorable Multiverse Analysis Reports' (EMARs) (Dragicevic P., Jansen Y., Sarma A., Kay M., Chevalier F., 2019) doi:10.1145/3290605.3300295 , a new approach to statistical reporting where readers of research papers can explore alternative analysis options by interacting with the paper itself.

multiverse is an R package that aims to make it easy to declare 'multiverse-style' analysis in R and R notebooks. The 'multiverse' package allows users to concisely and flexibly declare alternate ways of performing an analysis step in order to show how fragile or robust their findings are.

Details

'Multiverse style' analyses (Steegen 2016) is intended to highlight the robustness of an analysis to arbitrary decisions that are present in any data analysis. Considering all possible combinations of reasonable decisions that can be made at each step of an analysis, 'multiverse style' analysis can surface whether an outcome is an artifact of a particular idiosyncratic combination of analysis choices, or if it is robust against such arbitrary choices.

However, current tools do not support declaring 'multiverse' analysis easily, requiring users to declare custom control flows and multiple nested 'if-else' blocks. The multiverse package aims to simplify the process of composing 'multiverse' analysis using a flexible and concise syntax.

To get started with the multiverse package please refer to vignette("branch")`, vignette("conditions") and vignette("multiverse-in-rmd"). For example implementations of analysis using the multiverse package, see the case studies vignette("durante-multiverse-analysis") and vignette("hurricane").

References

Steegen, Sara, Francis Tuerlinckx, Andrew Gelman, and Wolf Vanpaemel. (2016). Increasing transparency through a multiverse analysis. Perspectives on Psychological Science, 11(5), 702-712. doi:10.1177/1745691616658637 .

See also

Author

Maintainer: Abhraneel Sarma abhraneel@u.northwestern.edu

Authors:

Matthew Kay mjskay@northwestern.edu

Other contributors:

Michael Moon moon@utstat.toronto.edu [contributor]

Mark Miller 725mrm@gmail.com [contributor]

Kyle Hwang kyle.sg.hwang@gmail.com [contributor]

Hadley Wickham hadley@posit.co [contributor]

Alex Kale kalea@uw.edu [contributor]

Nathan Taback nathan.taback@utoronto.ca [contributor]

Fanny Chevalier fanny@dgp.toronto.edu [contributor]

Jessica Hullman jessica.hullman@gmail.com [contributor]

Pierre Dragicevic pierre.dragicevic@inria.fr [contributor]

Yvonne Jansen yvonne.jansen@sorbonne-universite.fr [contributor]